src/spark/requirements.txt
delta-spark==2.4
pyspark==3.4